requirements.txt
Requirements：

python 3.7.13
numpy 1.16.5
scikit-learn 0.23.2
pytorch 1.10

streamlit==1.8.1
pandas==1.4.2
numpy~=1.24.3
scikit-learn~=1.2.2
openpyxl==3.0.9
plotly==5.7.0
kaleido==0.2.1
XlsxWriter==3.0.1
watchdog==2.1.7
Bio==1.3.3
lxml==4.6.5
matplotlib==3.3.4
matplotlib_venn==0.11.6
numpy==1.21.0
pandas==1.3.5
Pillow==9.0.1
py3Dmol==2.0.0.post2
pyfiglet==0.8.post1
rdkit_pypi==2021.9.4
requests==2.27.1
scipy~=1.10.1
seaborn==0.11.1
statannot==0.2.3
statsmodels==0.13.1
stmol==0.0.7
streamlit==1.8.1
tqdm==4.59.0

torch~=2.0.0